test(notebook-tools): consolidate extract_slidev_titles test duplication (#10066 tranche 4)
Delete the legacy scripts/tests/test_extract_slidev_titles.py shadow (29
tests, 220 lines) after porting its 10 unique tests into the canonical
scripts/notebook_tools/tests/test_extract_slidev_titles.py home.

Root cause: both files shared module name test_extract_slidev_titles, so
sys.modules cached the legacy (alphabetical) and the canon's 24 tests were
DEAD in CI whenever both ran. Verified empirically: canon-alone 34,
legacy-alone 29, both-together 58 = 29x2 (not 53) -> canon fully shadowed.
3rd tranche (of 4) with this collision mechanism -> it is the dominant
failure mode.

Per-test reconciliation (not count): TestExtract (6 integration tests of
the extract() pipeline the canon entirely LACKED) ported name-for-name
EXACT MATCH. 3 split_blocks edge cases (separator_with_spaces,
not_separator_with_text, single_separator) + 1 is_frontmatter case
(indented_yaml_not_detected) ported. Remaining legacy tests covered by
canon's own more-granular suite. This was the 2nd RISK file (legacy 29 >
canon 24 by count) where count-based deletion would have destroyed canon
coverage.

Effective live coverage 29 -> 34 (+5 net) + extract() integration tests
unified in the canonical home.

See #10066 (tranche 4 of 12). See #10055 (CI break that motivated #10066).
