+++ scripts/notebook_tools/tests/test_extract_slidev_titles.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent))
-from extract_slidev_titles import split_blocks, is_frontmatter, title_of
+from extract_slidev_titles import split_blocks, is_frontmatter, title_of, extract
 
 
 # --- split_blocks ---
@@ -52,6 +52,27 @@ class TestSplitBlocks:
         assert result[2] == ["# Slide 1", "Content"]
         assert result[3] == ["# Slide 2"]
 
+    # --- ported from deleted legacy scripts/tests/test_extract_slidev_titles.py
+    # shadow (#10066 consolidation): separator edge cases the canon lacked. ---
+
+    def test_separator_with_spaces(self):
+        """A separator line with only surrounding whitespace is still a separator
+        (split_blocks matches on ``line.strip() == '---'``)."""
+        lines = ["---", "  ---  ", "slide"]
+        result = split_blocks(lines)
+        assert len(result) == 3
+
+    def test_not_separator_with_text(self):
+        """``--- text`` (text after the dashes) is NOT a separator."""
+        lines = ["--- text", "content"]
+        result = split_blocks(lines)
+        assert len(result) == 1
+        assert result[0] == ["--- text", "content"]
+
+    def test_single_separator(self):
+        """A lone separator yields two empty blocks (minimal boundary case)."""
+        assert split_blocks(["---"]) == [[], []]
+
 
 # --- is_frontmatter ---
 
@@ -88,6 +109,15 @@ class TestIsFrontmatter:
     def test_multiline_yaml(self):
         block = ["layout: cover", "background: /images/bg.png"]
         assert is_frontmatter(block) is True
+
+    # --- ported from deleted legacy scripts/tests/test_extract_slidev_titles.py
+    # shadow (#10066 consolidation): the only frontmatter edge case the canon
+    # lacked — indented YAML must NOT be detected (YAML_KEY anchors at ^). ---
+
+    def test_indented_yaml_not_detected(self):
+        """An indented YAML line does not start with [A-Za-z_] at column 0, so it
+        is not treated as frontmatter (anchors the YAML_KEY regex)."""
+        assert is_frontmatter(["  layout: default"]) is False
 
 
 # --- title_of ---
@@ -128,6 +158,96 @@ class TestTitleOf:
     def test_h2_not_used_as_title(self):
         block = ["## Subtitle", "# Real Title"]
         assert title_of(block) == "Real Title"
+
+
+# --- extract (integration) ---
+#
+# Ported verbatim from the deleted legacy scripts/tests/test_extract_slidev_titles.py
+# shadow (#10066 consolidation). The canon previously tested only the pure
+# helpers (split_blocks / is_frontmatter / title_of) and never exercised the
+# ``extract()`` integration pipeline — reading a real .md file, dropping the
+# root frontmatter, interleaving per-slide frontmatter, and printing
+# ``idx: title`` lines. The module collision (both files shared the basename
+# ``test_extract_slidev_titles``) meant the canon's 24 helper tests were DEAD
+# in CI whenever both ran; deleting the legacy after porting these 6
+# integration tests resurrects the 24 AND unifies integration coverage in the
+# canonical home.
+
+
+class TestExtract:
+    """Integration tests for the full ``extract`` pipeline."""
+
+    def test_minimal_slidev(self, tmp_path, capsys):
+        md = tmp_path / "slides.md"
+        md.write_text(
+            "---\ntitle: Test\n---\n"
+            "# Slide 1\n\nContent\n"
+            "---\n"
+            "# Slide 2\n",
+            encoding="utf-8",
+        )
+        extract(md)
+        out = capsys.readouterr().out
+        assert "1: Slide 1" in out
+        assert "2: Slide 2" in out
+
+    def test_per_slide_frontmatter(self, tmp_path, capsys):
+        md = tmp_path / "slides.md"
+        md.write_text(
+            "---\ntitle: Test\n---\n"
+            "# Slide 1\n"
+            "---\n"
+            "layout: cover\n"
+            "---\n"
+            "# Slide 2\n",
+            encoding="utf-8",
+        )
+        extract(md)
+        out = capsys.readouterr().out
+        assert "1: Slide 1" in out
+        assert "2: Slide 2" in out
+
+    def test_no_h1_slide(self, tmp_path, capsys):
+        md = tmp_path / "slides.md"
+        md.write_text(
+            "---\ntitle: Test\n---\n"
+            "Just a paragraph\n",
+            encoding="utf-8",
+        )
+        extract(md)
+        out = capsys.readouterr().out
+        assert "1: [no-h1] Just a paragraph" in out
+
+    def test_empty_file(self, tmp_path, capsys):
+        md = tmp_path / "slides.md"
+        md.write_text("", encoding="utf-8")
+        extract(md)
+        out = capsys.readouterr().out
+        assert out == ""
+
+    def test_root_fm_only_produces_empty_slide(self, tmp_path, capsys):
+        """A trailing ``---`` produces one empty slide (split_blocks creates a
+        trailing block)."""
+        md = tmp_path / "slides.md"
+        md.write_text("---\ntitle: Test\n---\n", encoding="utf-8")
+        extract(md)
+        out = capsys.readouterr().out
+        assert "1: [empty]" in out
+
+    def test_three_slides(self, tmp_path, capsys):
+        md = tmp_path / "slides.md"
+        md.write_text(
+            "---\ntitle: Test\n---\n"
+            "# One\n---\n"
+            "# Two\n---\n"
+            "# Three\n",
+            encoding="utf-8",
+        )
+        extract(md)
+        out = capsys.readouterr().out
+        assert "1: One" in out
+        assert "2: Two" in out
+        assert "3: Three" in out
 
 
 if __name__ == "__main__":
